Allow poorly formatted SPK files (#40)
Some SPK files in the wild have poorly formatted contents, not naming names. These are interpolating a 12th order model with as low as 4 data points.
This change in the code will automatically reduce the order of the polynomial fit if not enough data is available.
When this happens, a warning is emitted to stderr.
Not the prettiest, but verbose enough that people should know the file is badly made.